pyproject.toml
[tool.poetry]
name = "fastapi-whistler-microservices"
version = "0.1.0"
description = ""
authors = ["itz-amethyst <>"]
license = "AGPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.4"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
python-slugify = "^8.0.4"
pymongo = "^4.8.0"
beanie = "^1.26.0"
captcha = "^0.6.0"
itsdangerous = "^2.2.0"
passlib = "^1.7.4"
structlog = "^24.4.0"
tenacity = "^8.5.0"
pyjwt = "^2.8.0"
emails = "^0.6"
aiosmtplib = "^3.0.1"
apscheduler = "^3.10.4"
prometheus-fastapi-instrumentator = "^7.0.0"
aiofiles = "^24.1.0"
aiohttp = "^3.10.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.1"
pytest-asyncio = "^0.23.8"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
